Fix tiles abruptly jumping after animation end
After the user releases the pointer, the tile animates to the target
final destination. We check whether an animation is still occurring by
checking whether the tile's current position is not equal to the tile's
final target position. The code erroneously had an debug statement that
skipped the check for almost all tiles, and this commit removes that
debug statement.