Separated React motion spring config into two conf
On tile pointer down, the spring physics controlling scaling, shadowing,
and move animation was the result of one spring config variable. Because
we want the move animation to complete faster with less precision, while
maintaining the slow animation and high precision of the scaling and
shadowing animations, there are now two spring physics configurations:
one for the move animation, and one for the scale and shadow animations.