Set static width to result to tiles shifting
Based on the size of result, tiles shift positions to keep the game
tiles and result centered to the screen. This is distracting for the
user and should be avoided. The result now has a static width of 5rem to
prevent repositioning. In the future, we should find a better solution
than a magic number of 5rem.