Fix double tap bug when triple tapping
When the user double taps, we check whether it is a double tap by
storing the last two times the tap occurred. In the case of a triple
tap, because we would always compare the last two stored tap times, a
triple click had the effect of two double clicks in a row. The proper
result should be that after a double click occurs, another double click
(two taps occurring within 300ms) should be required to trigger another
double click. This commit erases the stored tap times after a double tap
is successfuly made.