Fix operator hover effect for mobile and desktop
On mobile, tapping an operator would position the "invisible" cursor
over the operator, causing an indefinitely lasting hover effect. The
solution is to remove application of the hover effect from a standard
CSS :hover rule to a custom CSS class that can be added and removed via
JavaScript. Hooks for mouse enter, leave, and move, and hooks for touch
start, move, and end, are implemented to add and remove this CSS class
such that the hover class appears to be applied correctly on both mobile
and desktop devices.